refactor(#1610): device route util shared/ 이전 — #1604 backend 재사용 prereq
- src/features/route/utils/dijkstraRoute.ts → src/shared/utils/dijkstraRoute.ts
- src/features/route/utils/buildRouteGraph.ts → src/shared/utils/buildRouteGraph.ts
- 두 __tests__/ 파일도 동일 위치로 이전
- 내부 import path 업데이트 (../../../shared/* → ../*)
- backend Dijkstra infer(#1604)가 같은 알고리즘 재사용 — naive 복사 제거 prereq
